app-controller/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.27.0
pydantic==2.5.2
pyyaml==6.0.1
pynvml==11.5.0